Make dSFMT the default RNG.
- Fix overflow in energy with large N and D
- Code cleanup and rename variables

+++ src/state.h
@@ -7,7 +7,7 @@
 typedef struct
 {
     ising_spin *spin;
-    int magnetization;
+    int64_t magnetization;
 }
 state;
